Correct docker network issue in docker-compose.yaml
Updated Docker Compose file to fix an error which was preventing the docker container from seeing other containers and local llm's through ollama and LMstudio on linux.

With the old variation docker was creating a network specifically for this container which did not have access to localhost or other containers. This change forces the container to use the 'host' docker network allowing access to the bare metal llms.